requirements.txt
# 后端框架相关
fastapi>=0.68.0
uvicorn>=0.15.0
PyJWT>=2.10.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
pydantic-settings>=2.0.0
pydantic[email]
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5